// x-pack/platform/plugins/shared/dataset_quality/public/state_machines/dataset_quality_controller/src/types.d.ts
export type DefaultDatasetQualityControllerState = WithTableOptions & WithDataStreamStats & WithTotalDocs & WithDegradedDocs & WithFailedDocs & WithDatasets & WithFilters & WithAuthorizedDatasetTypes & WithNonAggregatableDatasets & Partial<WithIntegrations>;
type DefaultDatasetQualityStateContext = DefaultDatasetQualityControllerState;
export type DatasetQualityControllerTypeState = {
    value: 'initializing';
    context: DefaultDatasetQualityStateContext;
} | {
    value: 'initializationFailed';
    context: DefaultDatasetQualityStateContext;
} | {
    value: 'emptyState';
    context: DefaultDatasetQualityStateContext;
} | {
    value: 'main.stats.datasets.fetching';
    context: DefaultDatasetQualityStateContext;
} | {
    value: 'main.stats.datasets.loaded';
    context: DefaultDatasetQualityStateContext;
} | {
    value: 'main.stats.docsStats.fetching';
    context: DefaultDatasetQualityStateContext;
} | {
    value: 'main.stats.degradedDocs.fetching';
    context: DefaultDatasetQualityStateContext;
} | {
    value: 'main.stats.failedDocs.fetching';
    context: DefaultDatasetQualityStateContext;
} | {
    value: 'main.stats.nonAggregatableDatasets.fetching';
    context: DefaultDatasetQualityStateContext;
} | {
    value: 'main.integrations.fetching';
    context: DefaultDatasetQualityStateContext;
} | {
    value: 'main.nonAggregatableDatasets.fetching';
    context: DefaultDatasetQualityStateContext;
};
export type DatasetQualityControllerContext = DatasetQualityControllerTypeState['context'];
export type DatasetQualityControllerEvent = {
    type: 'UPDATE_TABLE_CRITERIA';
    dataset_criteria: TableCriteria<DatasetTableSortField>;
} | {
    type: 'UPDATE_INSIGHTS_TIME_RANGE';
    timeRange: TimeRangeConfig;
} | {
    type: 'TOGGLE_INACTIVE_DATASETS';
} | {
    type: 'TOGGLE_FULL_DATASET_NAMES';
} | {
    type: 'UPDATE_TIME_RANGE';
    timeRange: TimeRangeConfig;
} | {
    type: 'REFRESH_DATA';
} | {
    type: 'UPDATE_INTEGRATIONS';
    integrations: string[];
} | {
    type: 'UPDATE_NAMESPACES';
    namespaces: string[];
} | {
    type: 'UPDATE_QUALITIES';
    qualities: QualityIndicators[];
} | {
    type: 'UPDATE_QUERY';
    query: string;
} | {
    type: 'UPDATE_TYPES';
    types: DataStreamType[];
} | {
    type: 'UPDATE_FAILURE_STORE';
    dataStream: DataStreamStat;
} | {
    type: 'SAVE_TOTAL_DOCS_STATS';
    data: DataStreamDocsStat[];
    dataStreamType: DataStreamType;
} | {
    type: 'NOTIFY_TOTAL_DOCS_STATS_FAILED';
    error: Error;
};